Chore: Editor: Refactor editor package
