+++ .eslintignore
@@ -962,47 +962,46 @@ packages/editor/CodeMirror/editorCommands/duplicateLine.js
 packages/editor/CodeMirror/editorCommands/editorCommands.js
 packages/editor/CodeMirror/editorCommands/insertLineAfter.test.js
 packages/editor/CodeMirror/editorCommands/insertLineAfter.js
+packages/editor/CodeMirror/editorCommands/insertNewlineContinueMarkup.test.js
+packages/editor/CodeMirror/editorCommands/insertNewlineContinueMarkup.js
 packages/editor/CodeMirror/editorCommands/jumpToHash.test.js
 packages/editor/CodeMirror/editorCommands/jumpToHash.js
+packages/editor/CodeMirror/editorCommands/markdownCommands.bulletedVsChecklist.test.js
+packages/editor/CodeMirror/editorCommands/markdownCommands.test.js
+packages/editor/CodeMirror/editorCommands/markdownCommands.toggleList.test.js
+packages/editor/CodeMirror/editorCommands/markdownCommands.js
 packages/editor/CodeMirror/editorCommands/sortSelectedLines.test.js
 packages/editor/CodeMirror/editorCommands/sortSelectedLines.js
 packages/editor/CodeMirror/editorCommands/supportsCommand.js
+packages/editor/CodeMirror/extensions/biDirectionalTextExtension.js
+packages/editor/CodeMirror/extensions/keyUpHandlerExtension.js
+packages/editor/CodeMirror/extensions/markdownDecorationExtension.test.js
+packages/editor/CodeMirror/extensions/markdownDecorationExtension.js
+packages/editor/CodeMirror/extensions/markdownHighlightExtension.test.js
+packages/editor/CodeMirror/extensions/markdownHighlightExtension.js
+packages/editor/CodeMirror/extensions/markdownMathExtension.test.js
+packages/editor/CodeMirror/extensions/markdownMathExtension.js
+packages/editor/CodeMirror/extensions/overwriteModeExtension.test.js
+packages/editor/CodeMirror/extensions/overwriteModeExtension.js
+packages/editor/CodeMirror/extensions/searchExtension.js
+packages/editor/CodeMirror/extensions/selectedNoteIdExtension.js
 packages/editor/CodeMirror/getScrollFraction.js
-packages/editor/CodeMirror/markdown/MarkdownHighlightExtension.test.js
-packages/editor/CodeMirror/markdown/MarkdownHighlightExtension.js
-packages/editor/CodeMirror/markdown/MarkdownMathExtension.test.js
-packages/editor/CodeMirror/markdown/MarkdownMathExtension.js
-packages/editor/CodeMirror/markdown/codeBlockLanguages/allLanguages.js
-packages/editor/CodeMirror/markdown/codeBlockLanguages/defaultLanguage.js
-packages/editor/CodeMirror/markdown/codeBlockLanguages/lookUpLanguage.js
-packages/editor/CodeMirror/markdown/computeSelectionFormatting.test.js
-packages/editor/CodeMirror/markdown/computeSelectionFormatting.js
-packages/editor/CodeMirror/markdown/decoratorExtension.test.js
-packages/editor/CodeMirror/markdown/decoratorExtension.js
-packages/editor/CodeMirror/markdown/insertNewlineContinueMarkup.test.js
-packages/editor/CodeMirror/markdown/insertNewlineContinueMarkup.js
-packages/editor/CodeMirror/markdown/markdownCommands.bulletedVsChecklist.test.js
-packages/editor/CodeMirror/markdown/markdownCommands.test.js
-packages/editor/CodeMirror/markdown/markdownCommands.toggleList.test.js
-packages/editor/CodeMirror/markdown/markdownCommands.js
-packages/editor/CodeMirror/markdown/utils/renumberSelectedLists.test.js
-packages/editor/CodeMirror/markdown/utils/renumberSelectedLists.js
-packages/editor/CodeMirror/markdown/utils/stripBlockquote.js
 packages/editor/CodeMirror/pluginApi/PluginLoader.js
 packages/editor/CodeMirror/pluginApi/codeMirrorRequire.js
 packages/editor/CodeMirror/pluginApi/customEditorCompletion.test.js
 packages/editor/CodeMirror/pluginApi/customEditorCompletion.js
-packages/editor/CodeMirror/testUtil/createEditorControl.js
-packages/editor/CodeMirror/testUtil/createEditorSettings.js
-packages/editor/CodeMirror/testUtil/createTestEditor.js
-packages/editor/CodeMirror/testUtil/findNodesWithName.js
-packages/editor/CodeMirror/testUtil/forceFullParse.js
-packages/editor/CodeMirror/testUtil/loadLanguages.js
-packages/editor/CodeMirror/testUtil/pressReleaseKey.js
-packages/editor/CodeMirror/testUtil/typeText.js
+packages/editor/CodeMirror/testing/createEditorControl.js
+packages/editor/CodeMirror/testing/createEditorSettings.js
+packages/editor/CodeMirror/testing/createTestEditor.js
+packages/editor/CodeMirror/testing/findNodesWithName.js
+packages/editor/CodeMirror/testing/forceFullParse.js
+packages/editor/CodeMirror/testing/loadLanguages.js
+packages/editor/CodeMirror/testing/pressReleaseKey.js
+packages/editor/CodeMirror/testing/typeText.js
 packages/editor/CodeMirror/theme.js
-packages/editor/CodeMirror/utils/biDirectionalTextExtension.js
 packages/editor/CodeMirror/utils/formatting/RegionSpec.js
+packages/editor/CodeMirror/utils/formatting/computeSelectionFormatting.test.js
+packages/editor/CodeMirror/utils/formatting/computeSelectionFormatting.js
 packages/editor/CodeMirror/utils/formatting/findInlineMatch.test.js
 packages/editor/CodeMirror/utils/formatting/findInlineMatch.js
 packages/editor/CodeMirror/utils/formatting/isIndentationEquivalent.js
@@ -1021,11 +1020,12 @@ packages/editor/CodeMirror/utils/handleLinkEditRequests.js
 packages/editor/CodeMirror/utils/handlePasteEvent.js
 packages/editor/CodeMirror/utils/isCursorAtBeginning.js
 packages/editor/CodeMirror/utils/isInSyntaxNode.js
-packages/editor/CodeMirror/utils/keyUpHandlerExtension.js
-packages/editor/CodeMirror/utils/overwriteModeExtension.test.js
-packages/editor/CodeMirror/utils/overwriteModeExtension.js
-packages/editor/CodeMirror/utils/searchExtension.js
-packages/editor/CodeMirror/utils/selectedNoteIdExtension.js
+packages/editor/CodeMirror/utils/markdown/codeBlockLanguages/allLanguages.js
+packages/editor/CodeMirror/utils/markdown/codeBlockLanguages/defaultLanguage.js
+packages/editor/CodeMirror/utils/markdown/codeBlockLanguages/lookUpLanguage.js
+packages/editor/CodeMirror/utils/markdown/renumberSelectedLists.test.js
+packages/editor/CodeMirror/utils/markdown/renumberSelectedLists.js
+packages/editor/CodeMirror/utils/markdown/stripBlockquote.js
 packages/editor/CodeMirror/utils/setupVim.js
 packages/editor/SelectionFormatting.js
 packages/editor/events.js
