+++ packages/editor/CodeMirror/extensions/biDirectionalTextExtension.ts
