+++ packages/editor/CodeMirror/extensions/selectedNoteIdExtension.ts
