+++ packages/editor/CodeMirror/pluginApi/customEditorCompletion.test.ts
@@ -1,8 +1,8 @@
 import { EditorState, StateEffect } from '@codemirror/state';
-import createEditorControl from '../testUtil/createEditorControl';
+import createEditorControl from '../testing/createEditorControl';
 import { EditorView } from '@codemirror/view';
 import { completeFromList, completionStatus, currentCompletions } from '@codemirror/autocomplete';
-import typeText from '../testUtil/typeText';
+import typeText from '../testing/typeText';
 
 const waitForShownCompletionsToContain = (editor: EditorView, completionLabels: string[]) => {
 	return new Promise<string[]>(resolve => {
