+++ packages/editor/CodeMirror/testing/createEditorControl.ts
