+++ packages/editor/CodeMirror/testing/createEditorSettings.ts
