+++ packages/editor/CodeMirror/testing/forceFullParse.ts
