+++ packages/editor/CodeMirror/testing/typeText.ts
