+++ packages/editor/CodeMirror/utils/formatting/computeSelectionFormatting.test.ts
@@ -1,5 +1,5 @@
 import { EditorSelection } from '@codemirror/state';
-import createTestEditor from '../testUtil/createTestEditor';
+import createTestEditor from '../../testing/createTestEditor';
 import computeSelectionFormatting from './computeSelectionFormatting';
