+++ packages/editor/CodeMirror/utils/markdown/renumberSelectedLists.test.ts
@@ -1,5 +1,5 @@
 import { EditorSelection } from '@codemirror/state';
-import createTestEditor from '../../testUtil/createTestEditor';
+import createTestEditor from '../../testing/createTestEditor';
 import renumberSelectedLists from './renumberSelectedLists';
 
 describe('renumberSelectedLists', () => {
